+++ views/CHOICEPAGE1.handlebars
@@ -1,0 +1,25 @@
+<div class="row center" id="topScreen">
+    <h4 class="center">CHOICE PAGE 1</h4>
+    <p>WHAT DO YOU DO???</p>
+    
+    <a href="/pathA">
+    <button class="btn btn-success btn-lg">
+    <span class="glyphicon glyphicon-list-alt"></span>CHOICE1</button>
+    {{!-- THIS GOES TO PATH A --}}
+    </a>
+
+        <a href="/pathB">
+    <button class="btn btn-success btn-lg">
+    <span class="glyphicon glyphicon-list-alt"></span>CHOICE2</button>
+    {{!-- THIS GOES TO PATH B --}}
+    </a>
+
+        <a href="/pathC">
+    <button class="btn btn-success btn-lg">
+    <span class="glyphicon glyphicon-list-alt"></span>CHOICE3</button>
+    {{!-- THIS GOES TO PATH C --}}
+    </a>
+    
+    
+
+</div>
