+++ views/EXPO1.handlebars
@@ -1,0 +1,9 @@
+<div class="row center" id="topScreen">
+    <h4 class="center">EXPO PAGE 1</h4>
+    <a class="waves-effect waves-light btn-large internal-link">Start</a>
+    <P>YOU SEE SOMETHING IT HAS FEATURES</P>
+    <a href="/CHOICEPAGE1">
+        <button class="btn btn-success btn-lg">
+            <span class="glyphicon glyphicon-list-alt"></span>CHOICE1</button>
+    </a>
+</div>
