--- views/chapter1.handlebars
@@ -1,17 +1,0 @@
-<div class="row center" id="topScreen">
-    <h4 class="center">User Namefdjafjkldjakljfkdjaf</h4>
-    <a class="waves-effect waves-light btn-large internal-link">Start</a>
-    <a href="/userInputScreen">
-        <button class="btn btn-success btn-lg">
-            <span class="glyphicon glyphicon-list-alt"></span>START</button>
-    </a>
-    <form>
-        First name:
-        <input type="text" name="FirstName" value="Mickey">
-        <br> Last name:
-        <input type="text" name="LastName" value="Mouse">
-        <br>
-        <input type="submit" value="Submit">
-    </form>
-
-</div>
