+++ views/pathA.handlebars
@@ -1,0 +1,9 @@
+<div class="row center" id="topScreen">
+    <h4 class="center">PATH A</h4>
+    <a class="waves-effect waves-light btn-large internal-link">Start</a>
+    <P>YOU MADE CHOICE 1</P>
+    <a href="/pathA">
+        <button class="btn btn-success btn-lg">
+            <span class="glyphicon glyphicon-list-alt"></span>continue?</button>
+    </a>
+</div>
