+++ views/pathB.handlebars
@@ -1,0 +1,9 @@
+<div class="row center" id="topScreen">
+    <h4 class="center">PATH B</h4>
+    <a class="waves-effect waves-light btn-large internal-link">Start</a>
+    <P>YOU MADE CHOICE 2</P>
+    <a href="/pathB">
+        <button class="btn btn-success btn-lg">
+            <span class="glyphicon glyphicon-list-alt"></span>continue?</button>
+    </a>
+</div>
