+++ views/pathC.handlebars
@@ -1,0 +1,9 @@
+<div class="row center" id="topScreen">
+    <h4 class="center">PATH C</h4>
+    <a class="waves-effect waves-light btn-large internal-link">Start</a>
+    <P>YOU MADE CHOICE 3</P>
+    <a href="/pathC">
+        <button class="btn btn-success btn-lg">
+            <span class="glyphicon glyphicon-list-alt"></span>continue?</button>
+    </a>
+</div>
