+++ views/user.handlebars
@@ -1,10 +1,6 @@
 <div class="row center" id="topScreen">
     <h4 class="center">User Name Input Screen</h4>
     <a class="waves-effect waves-light btn-large internal-link">Start</a>
-    <a href="/userInputScreen">
-        <button class="btn btn-success btn-lg">
-            <span class="glyphicon glyphicon-list-alt"></span>START</button>
-    </a>
     <form>
         First name:
         <input type="text" name="FirstName" value="Mickey">
@@ -13,6 +9,9 @@
         <br>
         <input type="submit" value="Submit">
     </form>
-            <a href="/chapter1"><button class="btn btn-success btn-lg"><span class="glyphicon glyphicon-list-alt"></span>START</button></a>
+    <a href="/EXPO1">
+        <button class="btn btn-success btn-lg">
+            <span class="glyphicon glyphicon-list-alt"></span>START</button>
+    </a>
 
 </div>
